+++ cflib/wdial_get.c
@@ -31,7 +31,7 @@
  * Liefert zu einem Fensterhandle den Dialog oder NULL.
  */
 WDIALOG *
-get_wdial (int w_handle)
+get_wdial (short w_handle)
 {
 	WDIALOG *p;
 
@@ -40,7 +40,9 @@ get_wdial (int w_handle)
 	{
 		if (p->win_handle == w_handle)
 			return p;
+
 		p = p->next;
 	}
+
 	return NULL;
 }
