+++ cflib/wdial_uniconify.c
@@ -28,15 +28,15 @@
 
 
 void
-uniconify_wdial (WDIALOG * wd, int x, int y, int w, int h)
+uniconify_wdial (WDIALOG *wd, short x, short y, short w, short h)
 {
 	if (wd->mode & WD_ICON)
 	{
 		if (x == -1)
-			wind_set_grect (wd->win_handle, WF_UNICONIFY,
-					&wd->work);
+			wind_set_grect (wd->win_handle, WF_UNICONIFY, &wd->work);
 		else
 			wind_set (wd->win_handle, WF_UNICONIFY, x, y, w, h);
+
 		wind_get_grect (wd->win_handle, WF_WORKXYWH, &wd->work);
 		wd->mode &= ~WD_ICON;
 	}
